requirements.txt
flask 
tensorflow 
numpy
pillow
flask_cors
pandas
gunicorn